converting all Ehr resend integration tests into unit tests
